wireless: carl9170: Enable sniffer mode promisc flag to fix  injection
* Fix frame injection

The removal of the AR9170_MAC_SNIFFER_ENABLE_PROMISC flag to fix an issue
many years ago caused the AR9170 to not be able to pass probe response
packets with different MAC addresses back up to the driver. In general
operation, this doesn't matter, but in the case of packet injection with
aireplay-ng it is important. aireplay-ng specifically injects packets with
spoofed MAC addresses on the probe requests and looks for probe responses
back to those addresses. No other combination of filter flags seem to fix
this issue and so AR9170_MAC_SNIFFER_ENABLE is required to get these packets.

This was originally caused by commit e0509d3bdd7365d06c9bf570bf9f11 which
removed this flag in order to avoid spurious ack noise from the hardware.
In testing for this issue, keeping this flag but not restoring the
AR9170_MAC_RX_CTRL_ACK_IN_SNIFFER flag on the rc_ctrl seems to solve this
issue, at least with the most current firmware v1.9.9.

Change-Id: c3ccbbf9b006d55a17956eac723545018454c412
Signed-off-by: Steve deRosier <derosier@cal-sierra.com>
Signed-off-by: TogoFire <togofire@mailfence.com>
(cherry picked from commit 023cddda059c1023859843fa01e764a93eeb3daa)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>