Kbuild,lto: account for cross compiler when checking for LTO
The lto script was written with respect to x86 LTO, hence it relied
on direct vars for checking the binutils and other compiler tool
versions straight from host. Cross compilation was never accounted
for, which lead to checking of older binaries of the host, and hence
disabling LTO for the target architecture.

This patch accounts for the same and LTO cross compilation works as
intended.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>

(cherry picked from commit 921892189cae4b4a2097acb68a1fa8dc9b19ca5f)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>